[IMP]shopinvader_api_quotation: Refactor code to split sale and quotation schemas
