+++ shopinvader_api_quotation/schemas/__init__.py
@@ -1,1 +1,11 @@
-from . import sale
+from .sale import Sale
+from .quotation import (
+    QuotationState,
+    QuotationLines,
+    QuotationCreateRequest,
+    QuotationUpdateInput,
+    QuotationConfirmInput,
+    QuotationAddLineRequest,
+    QuotationUpdateLineRequest,
+    QuotationDeleteLineRequest,
+)
